+++ common/src/main/java/dev/terminalmc/chatnotify/config/IntermediaryConfigDeserializer.java
@@ -66,7 +66,7 @@ public class IntermediaryConfigDeserializer implements JsonDeserializer<Config> 
                     notifObj.getAsJsonArray("responseMessages")
                     .asList().stream().map(je2 -> new ResponseMessage(true, je2.getAsString(), false, 0)).toList());
 
-            notifications.add(new Notification(enabled, exclusionEnabled, responseEnabled,
+            notifications.add(new Notification(enabled, exclusionEnabled, responseEnabled, false,
                     sound, textStyle, new TitleText(), triggers, exclusionTriggers, responseMessages));
         }