+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Fixed style target string field overlapping delete button
 - Moved title text field to advanced options
+- Added advanced option to block messages via notification